refactor: App.css에 작성되었던 스타일 header,Nav,Footer로 분할
