--- src/components/App.css
@@ -1,136 +1,0 @@
-@import "../global.css";
-
-html {
-  height: 100%;
-}
-body {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-nav {
-  display: flex;
-  justify-content: center;
-  position: sticky;
-  top: 0;
-  width: 100%;
-  background-color: #edf7ff;
-}
-.App {
-  height: 100%;
-}
-
-.gnb {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  height: 6.3rem;
-  padding: 0 3.2rem;
-
-  @media (min-width: 768px) {
-    height: 9.4rem;
-    max-width: 86.3rem;
-  }
-
-  @media (min-width: 1200px) {
-    height: 9.4rem;
-    max-width: 192rem;
-    padding: 0 20rem;
-  }
-}
-.logo {
-  height: 1.6rem;
-
-  @media (min-width: 768px) {
-    height: 2.4rem;
-  }
-}
-.cta {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 3.7rem;
-  cursor: pointer;
-  background-image: linear-gradient(135deg, var(--primary) 0%, #6ae3fe 100%);
-  border-radius: 0.8rem;
-  color: #f5f5f5;
-  font-size: 1.4rem;
-  font-weight: 600;
-
-  @media (min-width: 768px) {
-    height: 5.4rem;
-    border-radius: 0.8rem;
-    font-size: 1.8rem;
-  }
-}
-
-.cta-short {
-  width: 8rem;
-
-  @media (min-width: 768px) {
-    width: 12.8rem;
-  }
-}
-
-.appHeader {
-  width: fit-content;
-  margin: 20px auto 60px;
-}
-
-footer {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  height: 16rem;
-  background-color: var(--black);
-}
-
-.footer-box {
-  display: grid;
-  justify-content: space-between;
-  grid-template:
-    "footer-links sns"
-    ". ." 1fr
-    "copyright .";
-  width: 100%;
-  padding: 3.2rem;
-
-  @media (min-width: 768px) {
-    display: flex;
-    justify-content: space-between;
-    grid-template: "copyright footer-links sns";
-    height: fit-content;
-    max-width: 192rem;
-    padding: 3.2rem 10.4rem 0;
-  }
-}
-
-.copyright {
-  grid-area: copyright;
-  color: #676767;
-  font-family: Arial;
-  font-size: 1.6rem;
-}
-
-.footer-links {
-  grid-area: footer-links;
-  display: flex;
-  column-gap: 3rem;
-  padding-right: 1.8rem;
-}
-
-.footer-link {
-  color: #cfcfcf;
-  font-family: Arial;
-  font-size: 1.6rem;
-}
-
-.sns {
-  grid-area: sns;
-  display: flex;
-  column-gap: 1.2rem;
-  height: 2rem;
-}
