+++ src/components/common/AppFooter.jsx
@@ -2,6 +2,7 @@ import twitter from "../../images/twitter.svg";
 import youtube from "../../images/youtube.svg";
 import instagram from "../../images/instagram.svg";
 import facebook from "../../images/facebook.svg";
+import "./AppFooter.css";
 const socialMediaLinks = [
   {
     href: "https://www.facebook.com/",
