+++ src/components/folder/CardControl.jsx
@@ -1,0 +1,21 @@
+export function CardControl() {
+  return (
+    <div className="CardControl">
+      <h2>유용한 글</h2>
+      <div>
+        <div>
+          <img alt="공유"></img>
+          <p>공유</p>
+        </div>
+        <div>
+          <img alt="이름변경"></img>
+          <p>이름 변경</p>
+        </div>
+        <div>
+          <img alt="삭제"></img>
+          <p>삭제</p>
+        </div>
+      </div>
+    </div>
+  );
+}
