+++ src/components/folder/FolderMain.jsx
@@ -1,14 +1,15 @@
 import Card from "../common/Cards";
 import { SearchBar } from "../common/SearchBar";
+import { CardControl } from "./CardControl";
 
 function Sorting() {}
-function UsefulArticles() {}
+
 export function FolderMain() {
   return (
     <>
       <SearchBar></SearchBar>
       <Sorting></Sorting>
-      <UsefulArticles></UsefulArticles>
+      <CardControl></CardControl>
       <Card></Card>
     </>
   );
