+++ src/components/shared/Main.css
@@ -1,0 +1,11 @@
+@import "../../global.css";
+
+html {
+  height: 100%;
+}
+body {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
